requirements.txt
bandit==1.6.2
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.10
django-crispy-forms==1.9.2
django-filter==2.3.0
django-import-export==2.3.0
django-simple-history==2.8.0
et-xmlfile==1.0.1
gitdb==4.0.5
GitPython==3.1.7
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
pbr==5.4.5
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
stevedore==3.2.0
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
